% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{block.positives.coverage}
\alias{block.positives.coverage}
\title{Check the quality of the blocking}
\usage{
block.positives.coverage(ground.truth, label.ground.truth, blocks,
  posclass)
}
\arguments{
\item{ground.truth}{A dataframe of the ground truth pairs}

\item{label.ground.truth}{The name of the column that contains the ground truth}

\item{blocks}{A dataframe of pairs produced from the blocking}

\item{posclass}{How is the positive class expressed (e.g. 1)}
}
\value{
coverage A value [0,1] showing the coverage of positives in the blocking.
}
\description{
Measures the coverage of the ground truth positives in the blocking
}
