% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{evaluate.skyex}
\alias{evaluate.skyex}
\title{Evaluate SkyEx algorithms}
\usage{
evaluate.skyex(prediction, labels, posclass)
}
\arguments{
\item{prediction}{A vector with the predicted classes}

\item{labels}{A vector of the actual classes}

\item{posclass}{How is the positive class expressed (e.g. 1)}
}
\description{
Calculates precision, recall and F-measure
}
