% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{plot.pairs2D}
\alias{plot.pairs2D}
\title{Plot SkyEx pairs 2D}
\usage{
\method{plot}{pairs2D}(data, sim1, sim2, prediction, labels, posclass,
  colTP, colTN, colFP, colFN, legend, leg.x, leg.y, leg.font, title, xlab,
  ylab)
}
\arguments{
\item{data}{A dataframe of pairs}

\item{sim1}{Name of the column in the x axis}

\item{sim2}{Name of the column in the y axis}

\item{prediction}{A vector of the predicted classes}

\item{labels}{A vector of the actual labels}

\item{posclass}{How the positive class is expressed (e.g. posclass=1)}

\item{colTP}{Color of the true positives}

\item{colTN}{Color of the true negatives}

\item{colFP}{Color of the false positives}

\item{colFN}{Color of the false negatives}

\item{legend}{A boolean argument if the legend should be present or not (e.g. legend=TRUE)}

\item{leg.x}{The x coordinate of the legend}

\item{leg.y}{The y coordinate of the legend}

\item{leg.font}{The font size of the legend}

\item{title}{The title of the plot}

\item{xlab}{Naming the x-axis}

\item{ylab}{Naming the y-axis}
}
\description{
Two dimentional plotting of the pairs and marking the true positives (TP), true negative (TN), false positives (FP), and false negatives (FN).
}
