% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{plot.skyexd.cutoffs}
\alias{plot.skyexd.cutoffs}
\title{Plot SkyEx-D cut-offs}
\usage{
\method{plot}{skyexd.cutoffs}(skyexd.obj, xlab, ylab)
}
\arguments{
\item{skyexd.object}{A skyexd object}
}
\description{
Plot the first derivative of the density function for different SkyEx-D cut-offs
}
