% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{plot.skyexd.smooth}
\alias{plot.skyexd.smooth}
\title{Plot SkyEx-D for different smoothing coefficients}
\usage{
\method{plot}{skyexd.smooth}(skyexd.obj, smooth.coefficient)
}
\arguments{
\item{smooth.coefficient}{A value for the smoothing coefficient}

\item{skyexd.object}{A skyexd object}
}
\description{
Plot the first derivative of the density function for different SkyEx-D cut-offs
}
