% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{plot.skyexf.cutoffs}
\alias{plot.skyexf.cutoffs}
\title{Plot SkyEx-F cut-offs}
\usage{
\method{plot}{skyexf.cutoffs}(skyexf.object, metric, xlab, ylab)
}
\arguments{
\item{skyexf.object}{A skyexf ovject}
}
\description{
Plot different SkyEx-F cut-offs and the corresponding F-measure
}
