% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{prefix.blocking}
\alias{prefix.blocking}
\title{Prefix blocking}
\usage{
prefix.blocking(data, column, prefix_size)
}
\arguments{
\item{data}{A dataframe of entities}

\item{column}{The column name of the attribute on which the prefix should be calculated}

\item{prefix_size}{The maximal number of characters for prefix blocking}
}
\value{
blocks A dataframe of pairs
}
\description{
Creates blocks of entities that have the same prefix. Returns the pairs
}
