% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{semantic.similarity}
\alias{semantic.similarity}
\title{Pairwise semantic similarity}
\usage{
semantic.similarity(data, column1, column2, pythonpath, method)
}
\arguments{
\item{data}{A dataframe of pairs}

\item{column1}{The column name containing the semantics for the first entity}

\item{column2}{The column name containing the semantics for the second entity}

\item{pythonpath}{The path where the python file is saved}

\item{method}{The method for the semantic similarity; choose between path (Path similarity), and wup (Wu&Palmer)}
}
\value{
sim A vector of semantic similarities
}
\description{
Compares the pairs pairwise regarding their semantic attributes. Returns a vector of semantic similarity
}
