% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{skyexd}
\alias{skyexd}
\title{SkyEx-D labeling}
\usage{
skyexd(data, p, simlist, smooth.coefficient, posclass, negclass)
}
\arguments{
\item{data}{A dataframe of pairs}

\item{p}{A preference function (e.g. p<-high("sim1")*high("sim2"))}

\item{simlist}{Similarities to be taken into consideration for the density function (e.g. simlist= c("sim1", "sim2"))}

\item{smooth.coefficient}{A smoothing coefficient}

\item{posclass}{How is the positive class expressed (e.g. 1)}

\item{negclass}{How is the negative class expressed (e.g. 0)}
}
\value{
od A skyexd object
}
\description{
Pair Labeling using SkyEx-D
}
