% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{skyexf}
\alias{skyexf}
\title{SkyEx-F labeling}
\usage{
skyexf(data, p, label, posclass, negclass)
}
\arguments{
\item{data}{A dataframe of pairs}

\item{p}{A preference function (e.g. p<-high("sim1")*high("sim2"))}

\item{label}{Name of the label column (e.g. "Class")}

\item{posclass}{How is the positive class expressed (e.g. 1)}

\item{negclass}{How is the negative class expressed (e.g. 0)}
}
\value{
of A skyexf object
}
\description{
Pair Labeling using SkyEx-F
}
