% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{spatial.blocking}
\alias{spatial.blocking}
\title{Spatial blocking}
\usage{
spatial.blocking(data, longitude, latitude, max_distance)
}
\arguments{
\item{data}{A dataframe of entities}

\item{longitude}{The column name that contains the longitudes}

\item{latitude}{The column name that contains the latitudes}

\item{max_distance}{The maximal distance in meters allowed in a block}
}
\value{
blocks A dataframe of pairs
}
\description{
Creates blocks of entities that near spatially. Returns the pairs
}
