% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{spatial.similarity}
\alias{spatial.similarity}
\title{Pairwise spatial similarity}
\usage{
spatial.similarity(data, lat1, long1, lat2, long2, max_distance)
}
\arguments{
\item{data}{A dataframe of pairs}

\item{lat1}{The column name of the latitude of the first entity}

\item{long1}{The column name of the longitude of the first entity}

\item{lat2}{The column name of the latitude of the second entity}

\item{long2}{The column name of the longitude of the second entity}

\item{max_distance}{The maximal distance allowed}
}
\value{
sim A vector of spatial similarities
}
\description{
Compares the pairs pairwise regarding a spatial attribute. Returns a vector of spatial similarity
}
