% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{suffix.blocking}
\alias{suffix.blocking}
\title{Suffix blocking}
\usage{
suffix.blocking(data, column, suffix_size)
}
\arguments{
\item{data}{A dataframe of entities}

\item{column}{The column name of the attribute on which the suffix should be calculated}

\item{suffix_size}{The maximal number of characters for suffix blocking}
}
\value{
blocks A dataframe of pairs
}
\description{
Creates blocks of entities that have the same suffix. Returns the pairs
}
