% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{text.similarity}
\alias{text.similarity}
\title{Pairwise textual similarity}
\usage{
\method{text}{similarity}(data, method, column1, column2)
}
\arguments{
\item{data}{A dataframe of pairs}

\item{method}{A method for the text similarity, choose among levenshtein, cosine, jaccard, jaro-winker}

\item{column1}{The first column name of the attribute that will be compared}

\item{column2}{The second column name of the attribute that will be compared}
}
\value{
sim A vector of text similarities
}
\description{
Compares the pairs pairwise regarding a textual attribute. Returns a vector of text similarity
}
