% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagetest.R
\name{textual.blocking}
\alias{textual.blocking}
\title{Textual blocking}
\usage{
textual.blocking(data, column, method, max_distance)
}
\arguments{
\item{data}{A dataframe of entities}

\item{column}{The column name of the attribute that will be considered for blocking}

\item{max_distance}{The maximal distance allowed in a block}

\item{label}{Method for textual blocking; choose among levenshtein, cosine, jaccard, jarowinker, qgram}
}
\value{
blocks A dataframe of pairs
}
\description{
Creates blocks of entities that have textual similarity. Returns the pairs
}
